feat: I've updated the musgravite library to version 1.0.1
