+++ _libs/musgravite/musgravite.hpp
@@ -3,8 +3,6 @@
 
 
 
-
-
 #include <iostream>
 #include <vector>
 #include <functional>
@@ -19,8 +17,6 @@
 
 
 
-
-
 #define RESET   "\033[0m"
 #define RED     "\033[31m"
 #define GREEN   "\033[32m"
@@ -32,28 +28,12 @@
 
 
 
-
-
-bool useColor = true;
-
-
-
-void disableColors() {
-    useColor = false;
-}
+extern bool useColor;
 
 
 
-void printColored(const std::string& message, const std::string& color = "") {
-    if (useColor) {
-        std::cout << color << message << RESET << std::endl;
-    } 
-    else {
-        std::cout << message << std::endl;
-    }
-}
-
-
+void disableColors();
+void printColored(const std::string& message, const std::string& color = "");
 
 
 
@@ -65,68 +45,10 @@ class Musgravite {
             std::function<void()> setup = nullptr,
             std::function<void()> teardown = nullptr,
             std::chrono::seconds timeout = std::chrono::seconds(30)
-        ) :
-            name(name),
-            func(func),
-            setup(setup),
-            teardown(teardown),
-            timeout(timeout)
-        {}
+        );
 
-
-        void run() const {
-            try {
-                if (setup) setup();
-
-                auto start = std::chrono::high_resolution_clock::now();
-                bool asyncTest = isAsyncTest(func);
-
-                if (asyncTest) {
-                    std::promise<void> promise;
-                    auto future = promise.get_future();
-
-                    std::thread asyncThread([&]() {
-                        try {
-                            func();
-                            promise.set_value();
-                        } catch (...) {
-                            promise.set_exception(std::current_exception());
-                        }
-                    });
-
-                    auto status = future.wait_for(timeout);
-
-                    if (status == std::future_status::timeout) {
-                        asyncThread.detach();
-                    
-                        throw std::runtime_error("Async test timed out.");
-                    } else {
-                        asyncThread.join();
-                    }
-                } else {
-                    func();
-                }
-
-                auto end = std::chrono::high_resolution_clock::now();
-
-                std::chrono::duration<double, std::milli> duration = end - start; // Duration in milliseconds
-                printColored("Test " + name + " passed in " + std::to_string(duration.count()) + " milliseconds.", GREEN);
-
-                if (teardown) teardown();
-            } catch (const std::exception& exception) {
-                printColored("Test " + name + " failed: " + exception.what(), RED);
-
-                if (teardown) teardown();
-
-                throw;
-            }
-        }
-
-
-        std::string getName() const {
-            return name;
-        }
-
+        void run() const;
+        std::string getName() const;
 
 
     private:
@@ -136,15 +58,8 @@ class Musgravite {
         std::function<void()> teardown;
         std::chrono::seconds timeout;
 
-
-        bool isAsyncTest(const std::function<void()>& f) const {
-            std::string funcName = name;
-
-            return funcName.size() > 5 && funcName.compare(funcName.size() - 5, 5, "Async") == 0;
-        }
+        bool isAsyncTest(const std::function<void()>& f) const;
 };
-
-
 
 
 
@@ -153,84 +68,18 @@ class MusgraviteRunner {
         void addTest(
             const std::string& suiteName,
             const std::string& testName,
-
             std::function<void()> func,
             std::function<void()> setup = nullptr,
             std::function<void()> teardown = nullptr,
-            std::chrono::seconds timeout = std::chrono::seconds(30))
-        {
-            if (suites.find(suiteName) == suites.end()) {
-                suites[suiteName] = std::vector<Musgravite>();
-            }
+            std::chrono::seconds timeout = std::chrono::seconds(30)
+        );
 
-            suites[suiteName].emplace_back(testName, func, setup, teardown, timeout);
-        }
-
-
-        void runAllTests() {
-            int totalPassed = 0;
-            int totalTests = 0;
-            double totalTime = 0.0;
-
-            for (const auto& suite : suites) {
-                int passed = 0;
-                double suiteTime = 0.0;
-
-                std::vector<std::string> failedTests;
-
-                printColored("Running suite: " + suite.first, CYAN);
-
-                for (auto& test : suite.second) {
-                    try {
-                        auto start = std::chrono::high_resolution_clock::now();
-
-                        test.run();
-
-                        auto end = std::chrono::high_resolution_clock::now();
-
-                        std::chrono::duration<double, std::milli> duration = end - start; // Duration in milliseconds
-
-                        suiteTime += duration.count();
-                        passed++;
-                    } catch (const std::exception& e) {
-                        failedTests.push_back(test.getName() + ": " + e.what());
-                    }
-                }
-
-                double averageTime = suite.second.empty() ? 0.0 : suiteTime / suite.second.size();
-                totalTime += suiteTime;
-
-                printColored(std::to_string(passed) + "/" + std::to_string(suite.second.size()) + " tests passed in suite " + suite.first, GREEN);
-                printColored("Suite " + suite.first + " total time: " + std::to_string(suiteTime) + " milliseconds.", BLUE);
-                printColored("Suite " + suite.first + " average time per test: " + std::to_string(averageTime) + " milliseconds.", RESET);
-
-                if (!failedTests.empty()) {
-                    printColored("Failed tests:", RED);
-
-                    for (const auto& fail : failedTests) {
-                        printColored(fail, RED);
-                    }
-                }
-
-                totalPassed += passed;
-                totalTests += suite.second.size();
-            }
-
-            double overallAverageTime = totalTests == 0 ? 0.0 : totalTime / totalTests;
-
-            printColored("Overall results:", BOLD);
-            printColored(std::to_string(totalPassed) + "/" + std::to_string(totalTests) + " tests passed across all suites.", GREEN);
-            printColored("Total time for all suites: " + std::to_string(totalTime) + " milliseconds.", BLUE);
-            printColored("Overall average time per test: " + std::to_string(overallAverageTime) + " milliseconds.", RESET);
-        }
-
+        void runAllTests();
 
 
     private:
         std::map<std::string, std::vector<Musgravite>> suites;
 };
-
-
 
 
 
@@ -252,16 +101,12 @@ class MusgraviteRunner {
 #define IGNORE_TEST(test) \
     printColored("Test " #test " ignored.", YELLOW);
 
-
-
-
 #define ASSERT_TRUE(cond) \
     if (!(cond)) { \
         std::stringstream ss; \
         ss << "Assertion failed: " << #cond << ", file " << __FILE__ << ", line " << __LINE__; \
         throw std::runtime_error(ss.str()); \
     }
-
 
 #define ASSERT_FALSE(cond) \
     if (cond) { \
@@ -270,14 +115,12 @@ class MusgraviteRunner {
         throw std::runtime_error(ss.str()); \
     }
 
-
 #define ASSERT_EQUAL(expected, actual) \
     if ((expected) != (actual)) { \
         std::stringstream ss; \
         ss << "Assertion failed: expected " << (expected) << ", got " << (actual) << ", file " << __FILE__ << ", line " << __LINE__; \
         throw std::runtime_error(ss.str()); \
     }
-
 
 #define ASSERT_NOT_EQUAL(expected, actual) \
     if ((expected) == (actual)) { \
@@ -286,14 +129,12 @@ class MusgraviteRunner {
         throw std::runtime_error(ss.str()); \
     }
 
-
 #define ASSERT_NEAR(expected, actual, tol) \
     if (std::fabs((expected) - (actual)) > (tol)) { \
         std::stringstream ss; \
         ss << "Assertion failed: expected near " << (expected) << ", got " << (actual) << ", with tolerance " << (tol) << ", file " << __FILE__ << ", line " << __LINE__; \
         throw std::runtime_error(ss.str()); \
     }
-
 
 #define ASSERT_THROW(statement, exception_type) \
     try { \
@@ -308,7 +149,6 @@ class MusgraviteRunner {
         throw std::runtime_error(ss.str()); \
     }
 
-
 #define ASSERT_NO_THROW(statement) \
     try { \
         statement; \
@@ -317,8 +157,6 @@ class MusgraviteRunner {
         ss << "Assertion failed: expected no exception, but an exception was thrown, file " << __FILE__ << ", line " << __LINE__; \
         throw std::runtime_error(ss.str()); \
     }
-
-
